code100/mp3bot/requirements.txt
aiogram==3.3.0
python-dotenv==1.0.0
aiofiles==23.2.1
aiohttp==3.9.0